Move types to seperate file
