+++ src/index.tsx
@@ -5,7 +5,6 @@ import {
   withReact,
   RenderElementProps,
   RenderLeafProps,
-  ReactEditor,
 } from 'slate-react';
 import {
   createEditor,
@@ -16,26 +15,24 @@ import {
   Location,
   Descendant,
 } from 'slate';
-import { withHistory, HistoryEditor } from 'slate-history';
+import { withHistory } from 'slate-history';
 import isHotKey from 'is-hotkey';
 import Toolbar from './components/Toolbar';
 import ImageBlock from './components/ImageBlock';
 import {
   basePlugin,
-  PluginsEditor,
   insertTab,
   softEnter,
   lists,
-  ListEditor,
   links,
-  LinkEditor,
   images,
-  ImageEditor,
   markdownShortcuts,
 } from './plugins';
 import { serialize, deserializeHtmlString } from './serializer';
 import { debounce } from 'lodash';
 import { compose } from 'lodash/fp';
+
+import { Elements, Mark } from './custom-types';
 
 interface Props {
   value?: string;
@@ -53,96 +50,6 @@ interface Props {
 }
 
 export const DEFAULT_BLOCK = 'paragraph';
-export type Next = () => unknown;
-
-export const MARKS = ['bold', 'italic', 'code', 'underline'] as const;
-export type Mark = typeof MARKS[number];
-export type Elements =
-  | 'h1'
-  | 'h2'
-  | 'h3'
-  | 'h4'
-  | 'h5'
-  | 'paragraph'
-  | 'ul_list'
-  | 'ol_list'
-  | 'list_item'
-  | 'code_block'
-  | 'link'
-  | 'figure'
-  | 'image'
-  | 'image_caption'
-  | 'quote';
-
-export type CustomText = { text: string; children: [] } & {
-  [key in Mark]?: boolean;
-};
-
-export type ImageElement = {
-  type: 'image';
-  src?: string;
-  objectUrl: string;
-  fileKey?: string;
-  children: [];
-} & { [key: string]: unknown };
-export type ImageCaptionElement = {
-  type: 'image_caption';
-  children: CustomText[];
-};
-export type FigureElement = {
-  type: 'figure';
-  children: (ImageElement | ImageCaptionElement)[];
-};
-export type TextElement = {
-  type: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'paragraph';
-  children: CustomText[];
-};
-type CodeBlockElement = {
-  type: 'code_block';
-  children: TextElement[];
-};
-export type LinkElement = {
-  type: 'link';
-  children: (TextElement | CustomText)[];
-  url: string;
-};
-type QuoteElement = {
-  type: 'quote';
-  children: CustomText[];
-};
-export type ListItemElement = {
-  type: 'list_item';
-  children: (TextElement | CustomText | LinkElement)[];
-};
-export type ListElement = {
-  type: 'ol_list' | 'ul_list';
-  children: (ListElement | ListItemElement)[];
-};
-
-type CustomElement =
-  | ListElement
-  | ListItemElement
-  | TextElement
-  | ImageElement
-  | ImageCaptionElement
-  | FigureElement
-  | QuoteElement
-  | CodeBlockElement
-  | LinkElement;
-
-declare module 'slate' {
-  interface CustomTypes {
-    Editor: BaseEditor &
-      ReactEditor &
-      HistoryEditor &
-      ListEditor &
-      LinkEditor &
-      ImageEditor &
-      PluginsEditor;
-    Element: CustomElement;
-    Text: CustomText;
-  }
-}
 
 /**
  *  Returns a function to be used for matching against node types
